Bitcoin mining project
First we create a new mix project by running the command 'mix new
project1' on the terminal. This creates a new mix project with the
respective lib and config folders along with the mix.exs file. Here, we
need to make the following changes to the mix.exs file- We add escript:
[main_module: MainServer]] in the definition of project in the mix file
so that the main module can be loaded successfully. In our project, the
main module is MainServer.
After saving the mix.exs file, we can run our project by ./project1
[args]
In our case, if the [args] is a number (e.g k = 3), then the Server is
mining bitcoins with that many leading zeroes (3 in this case). If the
[args] is an IP address, then it is the worker which is asking to
participate in the bitcoin mining process. The first call message to
the Server returns the value of leading zeroes to the worker. The
second call to the server sends an appended randomized string to the
worker, generated by the server. The worker mines bitcoins with these
randomized strings and sends it to the server using the 'cast' function
if the leading number of zeroes matches the value of k. This is carried
out in a loop and every time the worker calls the server to obtain a
randomized string.